refactor for taskcat building of dependency packages
