+++ functions/source/MLSharedLibraries/requirements.txt
@@ -1,0 +1,34 @@
+attrs==21.2.0
+black==21.9b0
+boto3==1.18.46
+botocore==1.21.46
+click==8.0.1
+dill==0.3.4
+google-pasta==0.2.0
+importlib-metadata==4.8.1
+jmespath==0.10.0
+multiprocess==0.70.12.2
+mypy-extensions==0.4.3
+numpy==1.21.2
+packaging==21.0
+pandas==1.1.5
+pathos==0.2.8
+pathspec==0.9.0
+platformdirs==2.4.0
+pox==0.3.0
+ppft==1.6.6.4
+protobuf==3.18.0
+protobuf3-to-dict==0.1.5
+pyparsing==2.4.7
+python-dateutil==2.8.2
+pytz==2021.1
+regex==2021.9.24
+s3transfer==0.5.0
+sagemaker==2.59.5
+six==1.16.0
+smdebug-rulesconfig==1.0.1
+tomli==1.2.1
+typed-ast==1.4.3
+typing-extensions==3.10.0.2
+urllib3==1.26.7
+zipp==3.6.0
