+++ functions/source/StartMLTrainingDeploymentFunction/requirements.txt
@@ -1,0 +1,3 @@
+boto3==1.18.53
+pandas==1.3.4
+sagemaker==2.68.0
