+++ functions/source/TransformationHandler/requirements.txt
@@ -1,0 +1,1 @@
+boto3==1.18.53
