+++ functions/source/UploadDeviceManifest/MCHP_manifest_signer.crt
@@ -1,0 +1,12 @@
+-----BEGIN CERTIFICATE-----
+MIIBxzCCAWygAwIBAgIQc6HyMjrkT2TOY17FAX7XVjAKBggqhkjOPQQDAjA8MSEw
+HwYDVQQKDBhNaWNyb2NoaXAgVGVjaG5vbG9neSBJbmMxFzAVBgNVBAMMDkxvZyBT
+aWduZXIgMDAyMB4XDTE5MDgxNTE5NDc1OVoXDTIwMDgxNTE5NDc1OVowPDEhMB8G
+A1UECgwYTWljcm9jaGlwIFRlY2hub2xvZ3kgSW5jMRcwFQYDVQQDDA5Mb2cgU2ln
+bmVyIDAwMjBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABLCLrgPlT3OezntD9lC2
+ShwUhlx07fiq/VETJ+ITUAwbgrPjB/Xi9GchLIM7FwZSUGOEqRA6KtH32XMpTGHK
+mCCjUDBOMB0GA1UdDgQWBBTxV4oZ3JTZ3zB6/r9XM0k4E6/5wDAfBgNVHSMEGDAW
+gBTxV4oZ3JTZ3zB6/r9XM0k4E6/5wDAMBgNVHRMBAf8EAjAAMAoGCCqGSM49BAMC
+A0kAMEYCIQDKHgctLnq/zNqfB+1v0KRhDVPvRf6Dimt8aW9WLS0NWAIhAJvUe3uJ
+pkMG4zpov9FCoj4G340idEadm7mVbAd5GOB9
+-----END CERTIFICATE-----
