--- functions/source/api/requirements.txt
@@ -1,21 +1,0 @@
-appdirs==1.4.4
-astroid==2.4.2
-boto3==1.14.20
-botocore==1.17.20
-certifi==2020.6.20
-cffi==1.14.0
-cryptography==2.9.2
-distlib==0.3.1
-ecdsa==0.15
-filelock==3.0.12
-isort==4.3.21
-jmespath==0.10.0
-lazy-object-proxy==1.4.3
-mccabe==0.6.1
-pipenv==2020.6.2
-pyasn1==0.4.8
-python-dateutil==2.8.1
-python-jose==3.1.0
-rsa==4.6
-s3transfer==0.3.3
-six==1.15.0
