+++ templates/connect-sense.template.yaml
@@ -299,7 +299,7 @@ Resources:
     Properties:
       CodeUri: 
         Bucket: !Ref QSS3BucketName
-        Key: !Sub ${QSS3KeyPrefix}functions/packages/UploadDeviceManifest.zip
+        Key: !Sub ${QSS3KeyPrefix}functions/packages/UploadDeviceManifest/lambda.zip
       Handler: api.upload_manifest
       Role: !GetAtt ProvisionApiLambdaRole.Arn
       Runtime: python3.8
@@ -320,7 +320,7 @@ Resources:
     Properties:
       CodeUri: 
           Bucket: !Ref QSS3BucketName
-          Key: !Sub ${QSS3KeyPrefix}functions/packages/ProcessAlerts.zip
+          Key: !Sub ${QSS3KeyPrefix}functions/packages/ProcessAlerts/lambda.zip
       Handler: alerts.process_alerts
       Role: !GetAtt ProcessAlertsLambdaRole.Arn
       Runtime: python3.8
